refactor: jpaRepository, RedisRepository 디렉토리 분리
EnableJpaRepositories 어노테이션 AuthApplication에 추가
EnableRedisRepositories 어노테이션 AuthApplication에 추가
