+++ src/main/java/com/fanfixiv/auth/AuthApplication.java
@@ -4,10 +4,14 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 
 @SpringBootApplication
 @EnableCaching
 @EnableJpaAuditing
+@EnableJpaRepositories(basePackages = "com.fanfixiv.auth.repository.jpa")
+@EnableRedisRepositories(basePackages = "com.fanfixiv.auth.repository.redis")
 public class AuthApplication {
 
   public static void main(String[] args) {
