+++ src/main/java/com/fanfixiv/auth/repository/jpa/ProfileRepository.java
@@ -1,8 +1,10 @@
-package com.fanfixiv.auth.repository;
+package com.fanfixiv.auth.repository.jpa;
 
 import com.fanfixiv.auth.entity.ProfileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface ProfileRepository extends JpaRepository<ProfileEntity, Long> {
   boolean existsByNickname(String nickname);
 }
