+++ src/main/java/com/fanfixiv/auth/repository/redis/RedisEmailRepository.java
@@ -1,8 +1,10 @@
-package com.fanfixiv.auth.repository;
+package com.fanfixiv.auth.repository.redis;
 
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 import com.fanfixiv.auth.dto.redis.RedisEmailAuthDto;
 
+@Repository
 public interface RedisEmailRepository extends CrudRepository<RedisEmailAuthDto, String> {
 }
