+++ src/main/java/com/fanfixiv/auth/repository/redis/RedisResetRepository.java
@@ -1,8 +1,10 @@
-package com.fanfixiv.auth.repository;
+package com.fanfixiv.auth.repository.redis;
 
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 import com.fanfixiv.auth.dto.redis.RedisResetDto;
 
+@Repository
 public interface RedisResetRepository extends CrudRepository<RedisResetDto, String> {
 }
