+++ src/main/java/com/fanfixiv/auth/service/CustomUserDetailService.java
@@ -2,7 +2,8 @@ package com.fanfixiv.auth.service;
 
 import com.fanfixiv.auth.details.User;
 import com.fanfixiv.auth.entity.UserEntity;
-import com.fanfixiv.auth.repository.UserRepository;
+import com.fanfixiv.auth.repository.jpa.UserRepository;
+
 import lombok.RequiredArgsConstructor;
 
 import javax.transaction.Transactional;
