+++ src/main/java/com/fanfixiv/auth/service/LoginService.java
@@ -6,8 +6,8 @@ import com.fanfixiv.auth.dto.login.LogoutResultDto;
 import com.fanfixiv.auth.dto.redis.RedisAuthDto;
 import com.fanfixiv.auth.entity.UserEntity;
 import com.fanfixiv.auth.interfaces.UserRoleEnum;
-import com.fanfixiv.auth.repository.RedisAuthRepository;
-import com.fanfixiv.auth.repository.UserRepository;
+import com.fanfixiv.auth.repository.jpa.UserRepository;
+import com.fanfixiv.auth.repository.redis.RedisAuthRepository;
 import com.fanfixiv.auth.utils.JwtTokenProvider;
 import com.google.common.net.HttpHeaders;
